Extract a generic appdomain isolation scope, which can later be re-used for test execution too.
