+++ machine.vstestadapter.sln
@@ -26,6 +26,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomDelegateSpec", "Sourc
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Machine.VSTestAdapter.VSIX", "Source\Machine.VSTestAdapter.VSIX\Machine.VSTestAdapter.VSIX.csproj", "{36354A22-5BDC-4314-AF72-5EBC4BBD2D4F}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{405D88EB-F3ED-4B4C-B758-4EA98692F96F}"
+	ProjectSection(SolutionItems) = preProject
+		appveyor.yml = appveyor.yml
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
